+++ app/src/main/AndroidManifest.xml
@@ -32,7 +32,6 @@
     </queries>
 
     <application
-        android:name=".App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -43,7 +42,7 @@
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".Presentation.ui.main.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -52,15 +51,15 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".SearchActivity"
+            android:name=".Presentation.ui.search.SearchActivity"
             android:exported="true" />
         <activity
-            android:name=".LibraryActivity"
+            android:name=".Presentation.ui.library.LibraryActivity"
             android:exported="true" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".Presentation.ui.settings.SettingsActivity"
             android:exported="true" />
-        <activity android:name=".PlayerActivity" />
+        <activity android:name=".Presentation.ui.player.PlayerActivity" />
 
     </application>
