+++ app/src/main/java/com/example/playlistmaker/Presentation/ui/library/LibraryActivity.kt
@@ -1,10 +1,10 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.Presentation.ui.library
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.example.playlistmaker.R
 
 class LibraryActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
