+++ app/src/main/java/com/example/playlistmaker/data/dto/Track.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.data.dto
 
 import java.io.Serializable
