+++ app/src/main/java/com/example/playlistmaker/data/repository/SearchHistory.kt
@@ -1,37 +1,33 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.data.repository
 
 import android.content.SharedPreferences
+import com.example.playlistmaker.data.dto.Track
+import com.example.playlistmaker.domain.repository.HistoryRepository
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 
-class SearchHistory(private val sharedPreferences: SharedPreferences) {
+class SearchHistory(private val sharedPreferences: SharedPreferences) : HistoryRepository {
+
     private val gson = Gson()
     private val key = "search_history"
 
-    fun getHistory(): List<Track> {
+    override fun getHistory(): List<Track> {
         val json = sharedPreferences.getString(key, null) ?: return emptyList()
         val type = object : TypeToken<List<Track>>() {}.type
         return gson.fromJson(json, type) ?: emptyList()
     }
 
-    fun addTrack(track: Track) {
+    override fun addTrack(track: Track) {
         val history = getHistory().toMutableList()
-
-        // Удаляем трек, если он уже есть в истории
         history.removeIf { it.trackName == track.trackName }
-
-        // Добавляем новый трек
         history.add(0, track)
-
-        // Ограничиваем историю до 10 треков
         if (history.size > 10) {
             history.removeAt(history.size - 1)
         }
-
         saveHistory(history)
     }
 
-    fun clearHistory() {
+    override fun clearHistory() {
         sharedPreferences.edit().remove(key).apply()
     }
