+++ app/src/main/java/com/example/playlistmaker/domain/interactors/HistoryInteractorImpl.kt
@@ -1,0 +1,19 @@
+package com.example.playlistmaker.domain.interactors
+
+import com.example.playlistmaker.domain.repository.HistoryRepository
+import com.example.playlistmaker.data.dto.Track
+
+class HistoryInteractorImpl(private val repository: HistoryRepository) : HistoryInteractor {
+
+    override fun getHistory(): List<Track> {
+        return repository.getHistory()
+    }
+
+    override fun addTrackToHistory(track: Track) {
+        repository.addTrack(track)
+    }
+
+    override fun clearHistory() {
+        repository.clearHistory()
+    }
+}
