+++ app/src/main/java/com/example/playlistmaker/domain/interactors/PlayerInteractor.kt
@@ -1,0 +1,10 @@
+package com.example.playlistmaker.domain.interactors
+
+interface PlayerInteractor {
+    fun preparePlayer(url: String, onPrepared: () -> Unit, onCompletion: () -> Unit)
+    fun play()
+    fun pause()
+    fun release()
+    fun isPlaying(): Boolean
+    fun getCurrentPosition(): Int
+}
