+++ app/src/main/java/com/example/playlistmaker/domain/interactors/PlayerInteractorImpl.kt
@@ -1,0 +1,37 @@
+package com.example.playlistmaker.domain.interactors
+
+import android.media.MediaPlayer
+
+class PlayerInteractorImpl : PlayerInteractor {
+    private var mediaPlayer: MediaPlayer? = null
+
+    override fun preparePlayer(url: String, onPrepared: () -> Unit, onCompletion: () -> Unit) {
+        mediaPlayer = MediaPlayer().apply {
+            setDataSource(url)
+            prepareAsync()
+            setOnPreparedListener { onPrepared() }
+            setOnCompletionListener { onCompletion() }
+        }
+    }
+
+    override fun play() {
+        mediaPlayer?.start()
+    }
+
+    override fun pause() {
+        mediaPlayer?.pause()
+    }
+
+    override fun release() {
+        mediaPlayer?.release()
+        mediaPlayer = null
+    }
+
+    override fun isPlaying(): Boolean {
+        return mediaPlayer?.isPlaying ?: false
+    }
+
+    override fun getCurrentPosition(): Int {
+        return mediaPlayer?.currentPosition ?: 0
+    }
+}
