+++ app/src/main/java/com/example/playlistmaker/domain/interactors/SettingsInteractor.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.domain.interactors
+
+interface SettingsInteractor {
+    fun isDarkThemeEnabled(): Boolean
+    fun setDarkThemeEnabled(enabled: Boolean)
+}
