+++ app/src/main/java/com/example/playlistmaker/domain/interactors/SettingsInteractorImpl.kt
@@ -1,0 +1,14 @@
+package com.example.playlistmaker.domain.interactors
+
+import com.example.playlistmaker.domain.repository.SettingsRepository
+
+class SettingsInteractorImpl(private val repository: SettingsRepository) : SettingsInteractor {
+
+    override fun isDarkThemeEnabled(): Boolean {
+        return repository.isDarkThemeEnabled()
+    }
+
+    override fun setDarkThemeEnabled(enabled: Boolean) {
+        repository.setDarkThemeEnabled(enabled)
+    }
+}
