+++ app/src/main/java/com/example/playlistmaker/domain/repository/SettingsRepository.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.domain.repository
+
+interface SettingsRepository {
+    fun isDarkThemeEnabled(): Boolean
+    fun setDarkThemeEnabled(enabled: Boolean)
+}
