+++ app/src/main/java/com/example/playlistmaker/domain/repository/TrackRepository.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.domain.repository
+
+import com.example.playlistmaker.data.dto.Track
+
+interface TrackRepository {
+    suspend fun searchTracks(query: String): List<Track>
+}
