+++ app/src/main/res/layout/activity_library.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SearchActivity">
+    tools:context=".Presentation.ui.search.SearchActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
